requirements.txt
# requirements.txt
flask
torch
torchvision
timm
Pillow